# Allow DateOffset addition with Series