… pandas/tseries/tests/test_timeseries.py