Addresses #10699 - allowing relative offsets like `DateOffset(years=1)` to to be added to a `Series` containing datetimes for element-wise addition.

Previously only non-relative offsets that could be mapped to a time_delta (`Tick`, `Day`, etc) worked.

Some performance numbers are below - `Tick` based offsets should be unchanged, a subset of offsets (`DateOffset`, `YearEnd/Begin`, `MonthEnd/Begin`, `QuarterEnd/Begin`, `Week`, `BusinessDay`) will be sped up significantly; and other offsets will still fallback to a slow path, but now will raise a performance warning.


### Master

    In [1]: idx = pd.date_range(start='20140101', freq='T', periods=100000)

    In [2]: %timeit idx + pd.offsets.Day()
    100 loops, best of 3: 5.72 ms per loop

    In [3]: %timeit idx + pd.Timedelta(1, 'D')
    100 loops, best of 3: 5.67 ms per loop

    In [4]: %timeit idx + pd.offsets.DateOffset(years=1, months=2, days=1)
    1 loops, best of 3: 1.79 s per loop

    In [5]: %timeit idx + pd.offsets.WeekOfMonth(week=2, weekday=4)
    1 loops, best of 3: 22.8 s per loop

### PR

    In [3]: %timeit idx + pd.offsets.Day()
    100 loops, best of 3: 5.8 ms per loop

    In [4]: %timeit idx + pd.Timedelta(1, 'D')
    100 loops, best of 3: 5.73 ms per loop

    In [5]: %timeit idx + pd.offsets.DateOffset(years=1, months=2, days=1)
    10 loops, best of 3: 86.9 ms per loop

    In [6]: %timeit idx + pd.offsets.WeekOfMonth(week=2, weekday=4)
    1 loops, best of 3: 23.4 s per loop
    c:\users\chris\documents\python-dev\pandas\pandas\tseries\index.py:695: PerformanceWarning:     Non-vectorized DateOffset being applied to Series or DatetimeIndex  PerformanceWarning)