@jreback 2015-08-04
This is a nice idea, but it wasn't originally implemented because it has massive perf issues.
However, a divide & conquer strategy can fix most of this. 

```
In [61]: np.random.seed(1234)

In [62]: s = Series(pd.to_timedelta(np.random.randint(0,10000000,size=10000),unit='s') + Timestamp('20130101'))

In [63]: pd.set_option('max_rows',10)

In [64]: np.random.seed(1234)

In [65]: s = Series(pd.to_timedelta(np.random.randint(0,10000000,size=10000),unit='s') + Timestamp('20130101'))

In [66]: s
Out[66]: 
0      2013-01-06 15:03:11
1      2013-03-20 00:58:59
2      2013-02-20 10:57:42
3      2013-03-13 04:40:21
4      2013-01-12 07:42:04
               ...        
9995   2013-02-17 07:22:39
9996   2013-04-03 04:57:11
9997   2013-04-16 10:06:38
9998   2013-02-15 15:28:46
9999   2013-03-05 12:00:39
dtype: datetime64[ns]
```

Any offset that has a ``_inc`` (e.g. <= Day), is simply convertible to a ``Timedelta``. (this *might* work for Week as well as its a fixed interval).

Note that your PR is equivalent to item by item apply.

```
In [78]: %timeit s.apply(lambda x: x+pd.offsets.Day())
1 loops, best of 3: 347 ms per loop

In [79]: %timeit s+Timedelta('1D')
1000 loops, best of 3: 610 us per loop
```

Conversions to > Day are a bit tricky. Note that you need to convert using ``pd.DatetimeIndex()`` because the obvious ``s.dt.to_period()`` round-trips thru ``object`` which defeats the purpose.
This would do a MonthEnd conversion. I think some of this could be make easier with some additonal methods (prob to ``DatetimeIndex``)

```
In [67]: i = pd.DatetimeIndex(s)

In [68]: base = (((i.to_period('M')+1).to_timestamp()).to_period('D')-1).to_timestamp()

In [69]: off = pd.TimedeltaIndex(i.asi8-i.to_period('D').to_timestamp().asi8)

In [70]: result = s.apply(lambda x: x+pd.offsets.MonthEnd())

In [71]: %timeit s.apply(lambda x: x+pd.offsets.MonthEnd())
1 loops, best of 3: 569 ms per loop
In [72]: result
Out[72]: 
0      2013-01-31 15:03:11
1      2013-03-31 00:58:59
2      2013-02-28 10:57:42
3      2013-03-31 04:40:21
4      2013-01-31 07:42:04
               ...        
9995   2013-02-28 07:22:39
9996   2013-04-30 04:57:11
9997   2013-04-30 10:06:38
9998   2013-02-28 15:28:46
9999   2013-03-31 12:00:39
dtype: datetime64[ns]

In [74]: Series(base+off)
Out[74]: 
0      2013-01-31 15:03:11
1      2013-03-31 00:58:59
2      2013-02-28 10:57:42
3      2013-03-31 04:40:21
4      2013-01-31 07:42:04
               ...        
9995   2013-02-28 07:22:39
9996   2013-04-30 04:57:11
9997   2013-04-30 10:06:38
9998   2013-02-28 15:28:46
9999   2013-03-31 12:00:39
dtype: datetime64[ns]

In [77]: %timeit Series((((i.to_period('M')+1).to_timestamp()).to_period('D')-1).to_timestamp() + pd.TimedeltaIndex(i.asi8-i.to_period('D').to_timestamp().asi8))
100 loops, best of 3: 7.74 ms per loop
```

Finally, you need to do Anchored offsets like ``BusinessDay``, ``W-THU`` via ``.apply`` (.e.g. I am not sure these can be vectorized, though maybe for something like ``W-THU`` you could do it).

You should emit a ``PerformanceWarning`` if you accept these last ones (or just raise).