@jreback 2015-08-04
Furthermore this would need tests for each of the offsets as well as subtracting.