@jreback 2015-08-04
this could be a property of a ``DateimeIndex``, it is equivalent of ``.time``, but returns a much more useful format ``pd.TimedeltaIndex(i.asi8-i.to_period('D').to_timestamp().asi8)``, maybe ``.timedelta`` (though not that descsriptive)