@chris-b1 2015-08-04
Thanks.  For those offsets that can be vectorized, where would that logic sit?  Something like a `vectorized_apply` method that `DateOffset.__add__` calls to if the other is array-like?