@jreback 2015-08-04
you can prob put it in the same place (but maybe put in a separate method in ``_Timeop``). Though some of the functions can put put in ``DatetimeIndex`` i think.