@chris-b1 2015-08-06
@jreback can you take a look at the latest updates  It might not be all the way there, but implements most of the obvious fastpaths I could see.

Instead of `.timedelta` like you showed  I added a `period_delta(freq)` method to `DatetimeIndex` which is the same idea, but handles other frequencies too ('M' and 'W' used).  Wasn't sure if this should be public.