@jreback 2015-08-06
@chris-b1 looks pretty good. Just a bit or reorg to put the code in the appropriate spots.