@jreback 2015-08-06
also if you'd add some vbenches would be great (and show the results in the very top of the PR).