@chris-b1 2015-08-07
@jreback  - see the update,  I think I've got most everything addressed, thanks for all the notes.

I wasn't sure what the `DateOffset` method that adds to a `DatetimeIndex` should be called - you suggested `period_delta` - I think that may be confusing where it doesn't return a `Timedelta`?  I ended up using `apply_index` but would be glad to change it to something better.

I'm on Windows and haven't able to get vbench working quite right - I did add some new benchmarks, and put timeits at the top of the PR.