@jreback 2015-08-08
looks pretty good. couple of code comments (mostly minor, you did nice job on moving things around!).
couple of doc additions.