@chris-b1 2015-08-08
@jreback  - here's another update - I addressed your comments, and expanded the tests, which highlighted a couple other changes needed.
  - I had broke adding offsets to `Timedelta`s
  - Handling normalized offsets
  - `month` kwarg in `YearEnd/Begin`; matching rollforwards when date is on `End/Begin` point

I also added vectorized paths for the `BusinessDay`, `Week`, `Quarter` offsets - same of approach of manipulating their `Period` equivalent.