@jreback 2015-08-11
@chris-b1 looking pretty good. couple of doc changes / test changes. Almost there!