@chris-b1 2015-08-12
@jreback thanks, latest has the changes to doc / testing from your comments.