@jreback 2015-08-13
small doc adds. pls squash. ping when green.